refactor(OnThisPage): Move to `Docs`
